requirements.txt
numpy>=1.24,<2.0
chroma-hnswlib>=0.7.6
matplotlib>=3.7
huggingface-hub>=0.20
